Fix issue paging
`githubDb.step` was misbehaving again, so I switched `cacheComments` to also
just use `execute`.

    00:15:51 web.1     | Assertion failed: (stmt), function EIO_Step, file ../src/statement.cc, line 579.

The breakage caused us to only every fetch the first page of issues because the
server never arrived at `pageRequest`'s finished callback.

We also weren't fetching the last page of issues thanks to a fencepost.

Finally, we would crash on some issues coming in `undefined`. I don't know what
exactly is going on, but ignoring those ones seems to work just fine. The
callback is wrapped in a `nextTick` to save the stack.